push testing changes

+++ app.js
@@ -11,11 +11,11 @@ client.connect();
 var XMLHttpRequest = require("xmlhttprequest").XMLHttpRequest;
 
 var request = new XMLHttpRequest();
-request.open("GET", "https://decapi.me/twitch/uptime?channel=nicnova_", true);
+//request.open("GET", "https://decapi.me/twitch/uptime?channel=nicnova_", true);
 
 //Connect to api for game info 
 var request2 = new XMLHttpRequest();
-request2.open("GET", "https://decapi.me/twitch/game/nicnova_", true);
+//request2.open("GET", "https://decapi.me/twitch/game/nicnova_", true);
 
 //put the txt file of forbidden words in an array
 var fs = require('fs');
@@ -24,6 +24,7 @@ try {
     var data = fs.readFileSync('forbiddenwords.txt', 'utf8');
 
     var forbiddenWords = data.split("\n");
+    //console.log(forbiddenWords);
 
 } catch(e) {
     console.log('Error:', e.stack);
@@ -44,14 +45,14 @@ client.on('message', (channel, tags, message, self) => {
         client.say(channel, `@${tags.username}, Please do not swear or use hateful language on this channel! Thank you. `);
     }
 
-    if(message.toLowerCase()== '!streamers') {
+    if(message.toLowerCase() === '!streamers') {
         client.say(channel, `@${tags.username}, Please don't ask to see other streamers. It's considered rude.`);
     }
 
     //Code to delete messages - update with list of terms? 
 
     for (var i = 0; i < forbiddenWords.length; i++) {
-      if(message.toLowerCase() == forbiddenWords[i]){
+      if(message.toLowerCase().includes(forbiddenWords[i]) && !message.toLowerCase().includes("hello")) {
         client.deletemessage(channel, tags.id)
         client.say(channel, `@${tags.username}, Please do not swear or use hateful language on this channel! Thank you. `);
         break;
@@ -61,46 +62,46 @@ client.on('message', (channel, tags, message, self) => {
     //code to check stream length
 
     if(message.toLowerCase()== '!uptime') {
-      request.onload = function() {
+      request.open("GET", "https://decapi.me/twitch/uptime?channel=nicnova_", true);
+      request.onload = function(e) {
 
-        if(request.status == 200)
-        {        
-          if (this.responseText.includes("offline")) {
-          client.say(channel, `@${tags.username}, ${this.responseText} right now.`);
-          }
-          else {
-            client.say(channel, `@${tags.username}, nicnova_ has been streaming for ${this.responseText}`);
-          }
-        
+      if (request.status === 200) {     
+        if (this.responseText.includes("offline")) {
+        client.say(channel, `@${tags.username}, ${this.responseText} right now.`);
         }
-        else
-        {
-            console.log(`Error occurred: Status: ${request.status}`);
+        else {
+          client.say(channel, `@${tags.username}, nicnova_ has been streaming for ${this.responseText}`);
         }
-      };
-      request.send();
+      }
+      else
+      {
+        console.log(`Error occurred: Status: ${request.status}`);
+      }
+    };
+    request.send();
     }
 
     //code to check what game the streamer is playing 
-    if(message.toLowerCase()== '!game') {
-      request2.onload = function() {
 
-        if(request2.status == 200)
+
+    if(message.toLowerCase()== '!game') {
+      request.open("GET", "https://decapi.me/twitch/game/nicnova_", true);
+      request.onload = function() {
+
+        if(request.status === 200)
         {        
           if (this.responseText == "") {
             client.say(channel, `@${tags.username}, nicnova_ is offline right now.`);
             }
             else {
               client.say(channel, `@${tags.username}, nicnova_ is streaming ${this.responseText}`);
-            }
-        
+            }  
         }
         else
         {
-            console.log(`Error occurred: Status: ${request2.status}`);
+           console.log(`Error occurred: Status: ${request.status}`);
         }
       };
-      request2.send();
+      request.send();
     }
-
   });
